# artifacts/api-server/src/routes/research.ts
import { and, desc, eq, sql } from "drizzle-orm";
import { Router, type IRouter, type RequestHandler } from "express";
import {
  ExecuteCompanyResearchParams,
  ExecuteCompanyResearchResponse,
  GetResearchEconomicsParams,
  GetResearchEconomicsResponse,
  ListResearchWorkspaceParams,
  ListResearchWorkspaceResponse,
  UpdateResearchBudgetBody,
  UpdateResearchBudgetParams,
  UpdateResearchBudgetResponse,
} from "@workspace/api-zod";
import {
  companiesTable,
  companyEvidenceTable,
  db,
  organizationMembersTable,
  projectCompaniesTable,
  projectsTable,
  researchJobsTable,
  researchQuestionsTable,
} from "@workspace/db";
import { executeResearchNow } from "../lib/research";
import { evaluateSignalsForCompany } from "../lib/signal-packs";
import { evaluateClustersForCompany } from "../lib/signal-clusters";
import { evaluateOpportunity } from "../lib/opportunity-engine";
import { generateWhyForOpportunity } from "../lib/opportunity-why";
import {
  getResearchEconomics,
  maximumResearchBudgetLimits,
  upsertResearchBudget,
} from "../lib/research-economics";
import { getAuthenticatedUserId, requireAuth, requireInternalAdmin } from "../middlewares/auth";
import { requireOrgRole } from "../lib/authz";

const router: IRouter = Router();
type AsyncHandler = (...args: Parameters<RequestHandler>) => Promise<void>;
const asyncRoute = (handler: AsyncHandler): RequestHandler =>
  (req, res, next) => void handler(req, res, next).catch(next);

async function authorizeProject(userId: string, projectId: string) {
  const [project] = await db.select().from(projectsTable)
    .where(eq(projectsTable.id, projectId)).limit(1);
  if (!project) return { status: 404 as const };
  const [membership] = await db.select({ id: organizationMembersTable.id })
    .from(organizationMembersTable)
    .where(and(
      eq(organizationMembersTable.organizationId, project.organizationId),
      eq(organizationMembersTable.userId, userId),
    )).limit(1);
  return membership ? { project } : { status: 403 as const };
}

/**
 * Server-side budget validation (C6): both caps must be present, whole-dollar
 * positive integers, no larger than the configured maximums, and the daily
 * cap may not exceed the monthly cap. Null ("unlimited") is rejected.
 */
export function validateResearchBudgetInput(
  input: { dailyBudget: number | null; monthlyBudget: number | null },
  limits = maximumResearchBudgetLimits(),
): { ok: true; dailyBudget: number; monthlyBudget: number } | { ok: false; error: string } {
  const { dailyBudget, monthlyBudget } = input;
  if (dailyBudget === null || monthlyBudget === null) {
    return { ok: false, error: "Daily and monthly research budgets are both required" };
  }
  for (const [label, value, max] of [
    ["Daily", dailyBudget, limits.dailyBudget],
    ["Monthly", monthlyBudget, limits.monthlyBudget],
  ] as const) {
    if (!Number.isInteger(value) || value <= 0) {
      return { ok: false, error: `${label} research budget must be a positive whole-dollar amount` };
    }
    if (value > max) {
      return { ok: false, error: `${label} research budget cannot exceed $${max.toFixed(0)}` };
    }
  }
  if (dailyBudget > monthlyBudget) {
    return { ok: false, error: "Daily research budget cannot exceed the monthly research budget" };
  }
  return { ok: true, dailyBudget, monthlyBudget };
}

function questionPayload(question: typeof researchQuestionsTable.$inferSelect | null) {
  return question ? {
    id: question.id,
    questionType: question.questionType,
    questionText: question.questionText,
    reason: question.reason,
    providerCapability: question.providerCapability,
    priority: question.priority,
    expectedInformationGain: question.expectedInformationGain,
    estimatedCost: question.estimatedCost,
    status: question.status,
    lastResultSummary: question.lastResultSummary,
    lastAttemptAt: question.lastAttemptAt,
    answeredAt: question.answeredAt,
  } : null;
}

function jobPayload(job: typeof researchJobsTable.$inferSelect | null) {
  return job ? {
    id: job.id,
    status: job.status,
    providerCapability: job.providerCapability,
    resultCount: job.resultCount,
    sourceCount: job.sourceCount,
    errorCode: job.errorCode,
    errorMessage: job.errorMessage,
    createdAt: job.createdAt,
    completedAt: job.completedAt,
  } : null;
}

router.get("/projects/:projectId/research", requireAuth, asyncRoute(async (req, res) => {
  const params = ListResearchWorkspaceParams.safeParse(req.params);
  if (!params.success) return void res.status(404).json({ error: "Project not found" });
  const access = await authorizeProject(getAuthenticatedUserId(res), params.data.projectId);
  if (!access.project) return void res.status(access.status).json({ error: access.status === 403 ? "Project access denied" : "Project not found" });
  const companies = await db.select({
    projectCompany: projectCompaniesTable,
    company: companiesTable,
  }).from(projectCompaniesTable)
    .innerJoin(companiesTable, eq(projectCompaniesTable.companyId, companiesTable.id))
    .where(eq(projectCompaniesTable.projectId, params.data.projectId))
    .orderBy(desc(projectCompaniesTable.updatedAt));
  const payload = await Promise.all(companies.map(async ({ projectCompany, company }) => {
    const [question] = await db.select().from(researchQuestionsTable)
      .where(and(eq(researchQuestionsTable.projectId, params.data.projectId), eq(researchQuestionsTable.companyId, company.id)))
      .orderBy(desc(researchQuestionsTable.createdAt)).limit(1);
    const [job] = await db.select().from(researchJobsTable)
      .where(and(eq(researchJobsTable.projectId, params.data.projectId), eq(researchJobsTable.companyId, company.id)))
      .orderBy(desc(researchJobsTable.createdAt)).limit(1);
    const [{ count }] = await db.select({ count: sql<number>`count(*)` })
      .from(companyEvidenceTable).where(eq(companyEvidenceTable.companyId, company.id));
    return {
      projectCompanyId: projectCompany.id,
      companyId: company.id,
      companyName: company.canonicalName,
      domain: company.domain,
      researchStatus: projectCompany.researchStatus,
      buyerRole: projectCompany.buyerRole,
      intelligenceStage: projectCompany.buyerRole === "UNKNOWN" ? "NEEDS_MINIMUM_INTELLIGENCE" : "ROLE_RESOLVED",
      nextAction: projectCompany.buyerRole === "POTENTIAL_BUYER" ? "Research now" : "Review company qualification",
      latestResearchAt: projectCompany.latestResearchAt,
      evidenceCount: Number(count),
      question: questionPayload(question ?? null),
      job: jobPayload(job ?? null),
    };
  }));
  res.json(ListResearchWorkspaceResponse.parse(payload));
}));

// Spend, projections and dollar budgets are internal: admins only (26 Sep 2026).
router.get("/projects/:projectId/research/economics", requireAuth, requireInternalAdmin, asyncRoute(async (req, res) => {
  const params = GetResearchEconomicsParams.safeParse(req.params);
  if (!params.success) return void res.status(404).json({ error: "Project not found" });
  const access = await authorizeProject(getAuthenticatedUserId(res), params.data.projectId);
  if (!access.project) return void res.status(access.status).json({ error: access.status === 403 ? "Project access denied" : "Project not found" });
  const summary = await getResearchEconomics(params.data.projectId);
  res.json(GetResearchEconomicsResponse.parse(summary));
}));

router.put("/projects/:projectId/research/budget", requireAuth, requireInternalAdmin, asyncRoute(async (req, res) => {
  const params = UpdateResearchBudgetParams.safeParse(req.params);
  const body = UpdateResearchBudgetBody.safeParse(req.body);
  if (!params.success || !body.success) return void res.status(400).json({ error: "Invalid research budget" });
  const userId = getAuthenticatedUserId(res);
  const access = await authorizeProject(userId, params.data.projectId);
  if (!access.project) return void res.status(access.status).json({ error: access.status === 403 ? "Project access denied" : "Project not found" });
  if (!(await requireOrgRole(res, userId, access.project.organizationId))) return;
  const validation = validateResearchBudgetInput(body.data);
  if (!validation.ok) return void res.status(400).json({ error: validation.error });
  const budget = await upsertResearchBudget({
    organizationId: access.project.organizationId,
    projectId: params.data.projectId,
    createdBy: userId,
    dailyBudget: validation.dailyBudget,
    monthlyBudget: validation.monthlyBudget,
    currency: body.data.currency,
  });
  res.json(UpdateResearchBudgetResponse.parse({
    dailyBudget: budget.dailyBudget,
    monthlyBudget: budget.monthlyBudget,
    currency: budget.currency,
    updatedAt: budget.updatedAt,
  }));
}));

router.post("/projects/:projectId/companies/:projectCompanyId/research", requireAuth, asyncRoute(async (req, res) => {
  const params = ExecuteCompanyResearchParams.safeParse(req.params);
  if (!params.success) return void res.status(404).json({ error: "Project company not found" });
  const access = await authorizeProject(getAuthenticatedUserId(res), params.data.projectId);
  if (!access.project) return void res.status(access.status).json({ error: access.status === 403 ? "Project access denied" : "Project not found" });
  const result = await executeResearchNow({
    projectId: params.data.projectId,
    projectCompanyId: params.data.projectCompanyId,
    organizationId: access.project.organizationId,
    userId: getAuthenticatedUserId(res),
  });
  if ("stopped" in result) {
    res.json(ExecuteCompanyResearchResponse.parse({
      stopped: true,
      reason: result.reason,
      buyerRole: result.buyerRole ?? "UNKNOWN",
      intelligenceStage: result.intelligenceStage ?? "NOT_STARTED",
      stopCode: result.stopCode ?? "STOPPED",
      progress: result.progress ?? "buyer_research_gated",
      nextAction: result.nextAction ?? "Review the company role.",
      question: null,
      job: null,
      evidenceCount: 0,
      factProposalCount: 0,
      factRejectionCount: 0,
      resultStatus: "STOPPED",
    }));
    return;
  }
  if (result.resultStatus === "SUCCEEDED" && result.evidenceCount > 0) {
    const [row] = await db.select({
      companyId: projectCompaniesTable.companyId,
    }).from(projectCompaniesTable).where(and(
      eq(projectCompaniesTable.id, params.data.projectCompanyId),
      eq(projectCompaniesTable.projectId, params.data.projectId),
    )).limit(1);
    if (row) {
      await evaluateSignalsForCompany({
        organizationId: access.project.organizationId,
        projectId: params.data.projectId,
        companyId: row.companyId,
      });
      await evaluateClustersForCompany({
        organizationId: access.project.organizationId,
        projectId: params.data.projectId,
        companyId: row.companyId,
      });
      const opportunity = await evaluateOpportunity({
        organizationId: access.project.organizationId,
        projectId: params.data.projectId,
        projectCompanyId: params.data.projectCompanyId,
        userId: getAuthenticatedUserId(res),
      });
      await generateWhyForOpportunity(opportunity.opportunity.id, params.data.projectId);
    }
  }
  res.json(ExecuteCompanyResearchResponse.parse({
    stopped: false,
    reason: null,
    buyerRole: "POTENTIAL_BUYER",
    intelligenceStage: "BUYER_RESEARCH",
    stopCode: null,
    progress: "buyer_research_complete",
    nextAction: "Review the resulting evidence.",
    question: questionPayload(result.question),
    job: jobPayload(result.job),
    evidenceCount: result.evidenceCount,
    factProposalCount: result.factProposalCount,
    factRejectionCount: result.factRejectionCount,
    resultStatus: result.resultStatus,
  }));
}));

export default router;