# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{projectId}/research:
    get:
      operationId: listResearchWorkspace
      tags: [research]
      summary: List bounded research state for project companies
      security: [{ clerkSession: [] }]
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Research workspace
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ResearchWorkspaceCompany" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /projects/{projectId}/research/economics:
    get:
      operationId: getResearchEconomics
      tags: [research]
      summary: Get project research spend, projection, and unit economics (internal admins only)
      security: [{ clerkSession: [] }]
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Research economics summary
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ResearchEconomicsSummary" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /projects/{projectId}/research/budget:
    put:
      operationId: updateResearchBudget
      tags: [research]
      summary: Configure optional daily and monthly project research budgets (internal admins only)
      security: [{ clerkSession: [] }]
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ResearchBudgetInput" }
      responses:
        "200":
          description: Updated research budget
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ResearchBudget" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /projects/{projectId}/companies/{projectCompanyId}/research:
    post:
      operationId: executeCompanyResearch
      tags: [research]
      summary: Execute the highest-value due research question
      security: [{ clerkSession: [] }]
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/ProjectCompanyId"
      responses:
        "200":
          description: Research execution result
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ResearchExecutionResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  schemas:
...
    ResearchQuestion:
      type: object
      properties:
        id: { type: string }
        questionType: { type: string }
        questionText: { type: string }
        reason: { type: string }
        providerCapability: { type: string }
        priority: { type: number }
        expectedInformationGain: { type: number }
        estimatedCost: { type: number }
        status: { type: string }
        lastResultSummary: { type: ["string", "null"] }
        lastAttemptAt: { type: ["string", "null"], format: date-time }
        answeredAt: { type: ["string", "null"], format: date-time }
      required: [id, questionType, questionText, reason, providerCapability, priority, expectedInformationGain, estimatedCost, status, lastResultSummary, lastAttemptAt, answeredAt]
    ResearchJob:
      type: object
      properties:
        id: { type: string }
        status: { type: string }
        providerCapability: { type: string }
        resultCount: { type: number }
        sourceCount: { type: number }
        errorCode: { type: ["string", "null"] }
        errorMessage: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        completedAt: { type: ["string", "null"], format: date-time }
      required: [id, status, providerCapability, resultCount, sourceCount, errorCode, errorMessage, createdAt, completedAt]
    ResearchWorkspaceCompany:
      type: object
      properties:
        projectCompanyId: { type: string }
        companyId: { type: string }
        companyName: { type: string }
        domain: { type: ["string", "null"] }
        researchStatus: { type: string }
        buyerRole: { type: string, enum: [POTENTIAL_BUYER, SELLER_COMPETITOR, ADJACENT_VENDOR, PARTNER_POSSIBLE, UNKNOWN] }
        intelligenceStage: { type: string }
        nextAction: { type: string }
        latestResearchAt: { type: ["string", "null"], format: date-time }
        evidenceCount: { type: number }
        question: { oneOf: [{ $ref: "#/components/schemas/ResearchQuestion" }, { type: "null" }] }
        job: { oneOf: [{ $ref: "#/components/schemas/ResearchJob" }, { type: "null" }] }
      required: [projectCompanyId, companyId, companyName, domain, researchStatus, buyerRole, intelligenceStage, nextAction, latestResearchAt, evidenceCount, question, job]
    ResearchExecutionResponse:
      type: object
      properties:
        stopped: { type: boolean }
        reason: { type: ["string", "null"] }
        buyerRole: { type: string, enum: [POTENTIAL_BUYER, SELLER_COMPETITOR, ADJACENT_VENDOR, PARTNER_POSSIBLE, UNKNOWN] }
        intelligenceStage: { type: string }
        stopCode: { type: ["string", "null"] }
        progress: { type: string }
        nextAction: { type: string }
        question: { oneOf: [{ $ref: "#/components/schemas/ResearchQuestion" }, { type: "null" }] }
        job: { oneOf: [{ $ref: "#/components/schemas/ResearchJob" }, { type: "null" }] }
        evidenceCount: { type: number }
        factProposalCount: { type: number }
        factRejectionCount: { type: number }
        resultStatus: { type: string }
      required: [stopped, reason, buyerRole, intelligenceStage, stopCode, progress, nextAction, question, job, evidenceCount, factProposalCount, factRejectionCount, resultStatus]
    ResearchBudgetInput:
      type: object
      properties:
        dailyBudget: { type: ["number", "null"], minimum: 0 }
        monthlyBudget: { type: ["number", "null"], minimum: 0 }
        currency: { type: string, enum: [USD], default: USD }
      required: [dailyBudget, monthlyBudget, currency]
    ResearchBudget:
      type: object
      properties:
        dailyBudget: { type: ["number", "null"] }
        monthlyBudget: { type: ["number", "null"] }
        currency: { type: string }
        updatedAt: { type: string, format: date-time }
      required: [dailyBudget, monthlyBudget, currency, updatedAt]
    ResearchEconomicsSummary:
      type: object
      properties:
        currency: { type: string }
        dailyBudget: { type: ["number", "null"] }
        monthlyBudget: { type: ["number", "null"] }
        spendToday: { type: number }
        spendThisMonth: { type: number }
        projectedMonthSpend: { type: number }
        unknownCostRequestsThisMonth: { type: number }
        requestsThisMonth: { type: number }
        successfulRequestsThisMonth: { type: number }
        blockedRequestsThisMonth: { type: number }
        costPerCompanyResearched: { type: ["number", "null"] }
        costPerOpportunityIdentified: { type: ["number", "null"] }
        costPerBuyerFound: { type: ["number", "null"] }
        companiesResearchedThisMonth: { type: number }
        opportunitiesIdentified: { type: number }
        buyersFound: { type: number }
      required: [currency, dailyBudget, monthlyBudget, spendToday, spendThisMonth, projectedMonthSpend, unknownCostRequestsThisMonth, requestsThisMonth, successfulRequestsThisMonth, blockedRequestsThisMonth, costPerCompanyResearched, costPerOpportunityIdentified, costPerBuyerFound, companiesResearchedThisMonth, opportunitiesIdentified, buyersFound]
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
  parameters:
...
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
...
    ProjectCompanyId:
      name: projectCompanyId
      in: path
      required: true
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Authenticated but not allowed to access this resource
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  securitySchemes:
    clerkSession:
      type: apiKey
      in: cookie
      name: __session